add a modal for migrations
